[refactor] Refactor v1.6.2
Initial refactor of prototype before plugins implementation.
This was done as plugins are cancelled, however there were many features
built during that time that were desirable. This reimplements some of
those features.

Changes
- Removed app plugin code (Discord, VTS, Twitch, etc)
- Converted models into operations
- Added operation streaming
- Updated response pipeline
- Updated and extended JAIson application layer interface
- Reformat logs
- Reformat configuration
- Adjust CLI arguments
- Updated observer classes to use message queues
- Added sequential job queueing for consistent and undisturbed state
  during response generation
- Added external process management for starting API servers it depends
   on
- Added integration with KoboldCPP
- Added sentence chunking
- Added compatibility mode
- Integrated RVC package

Project is still in WIP state. It is not ready to run and has yet to be
tested.

Signed-off-by: limitcantcode <limit@limit-cc.com>