Prototype Refactor (#20)
This is a complete revamp and polishing of the prototype prior to v2.0.0. This was motivated by the choice to consolidate server code into one project while keeping design patterns simple, maintainable, and extensible. This moves all front-facing application code into their own projects, but for everything on the server side, the same features have been polished and new features/integrations have been added. This will serve as the new base moving forward, and v2.0.0 will be deprecated.

Signed-off-by: limitcantcode <limit@limit-cc.com>